Redo auth plugins
There are a number of changes here to better align OSC with how the
other clients are expecting to use the keystoneclient auth plugins.

Firstly I think it is a bad idea to Combine all the options from every
available plugin into the --help. There are plugins recently for example
for federation that have a bunch of possible options that are going to
polute this space with options that are not relevant for most people.

Instead of this define a default plugin that handles the options that
were initially available to OSC. Additional options were added in a
recent change however they have not yet been in a release. If we want to
extend the available options for OSC then those options can be added to
the default plugin.

Change-Id: I759cfb2e3989c1abd5cd49826eb136a60427239e